msm: kgsl: Tweek LLM sleep/wake algorithm
In the LLM sleep sequence, the IDLE_FULL_LM bit (0) needs to be set
to force the children to sleep.  Also, after the first child is
put to sleep, we need to wait for the idle acknowledgment before
taking the next child down.

In the wake sequence, poll until WAKEUP_ACK is 1 *and* IDLE_FULL_ACK
is 0 to ensure that the wake sequence was successful.

CRs-Fixed: 970270
Change-Id: Ic0dedbadfca1e0882d84965d634166f921f1e630
Signed-off-by: Jordan Crouse <jcrouse@codeaurora.org>